fix(spring): split and reorder queue metrics bean definitions
- splitting QueueMonitor out breaks the QueueMetricsPublisher <-> MonitorableQueue circular dependency
- @ConditionalOnMissingBean depends on the order things are defined, so NoopEventPublisher needs to be
  defined after QueueMetricsPublisher to avoid a conflict
